requirements.txt
langchain
langchain_community
langchain_google_genai
protobuf
PyPDF2
streamlit
